~ Tous les degats et les santes sont maintenant gérés en long
+ Astuce secrete (debug) pour gagner de l'argent... je vous expliquerai ;)

~ Amélioration de l'affichage de la sante des creatures (2345 = 2.3K, 40'567'890 = 40.5M, 999'876'543'212 = 999.8G)

~ J'ai remis les Nuages (2 vagues de volantes : Nuage et Boo)

PS : j'ai joué un bon moment et je n'ai pas revu le bug des animations folles :)

+++ trunk/src/models/attaques/BouleDeFeu.java
@@ -52,7 +52,7 @@ public class BouleDeFeu extends Attaque
      * @param cible la creature visee
      */
     public BouleDeFeu(Terrain terrain, Tour attaquant, Creature cible, 
-                      int degats, double rayonImpact)
+                      long degats, double rayonImpact)
     {
         super((int) attaquant.getCenterX(),(int) attaquant.getCenterY(), terrain, attaquant, cible);
