+++ trunk/src/models/creatures/GrandeFlame.java
@@ -30,7 +30,7 @@ public class GrandeFlame extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public GrandeFlame(int santeMax, int nbPiecesDOr, double vitesse)
+	public GrandeFlame(long santeMax, int nbPiecesDOr, double vitesse)
 	{
 		this(0, 0, santeMax, nbPiecesDOr,vitesse);
 	}
@@ -44,7 +44,7 @@ public class GrandeFlame extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public GrandeFlame(int x, int y, int santeMax, int nbPiecesDOr, double vitesse)
+	public GrandeFlame(int x, int y, long santeMax, int nbPiecesDOr, double vitesse)
 	{
 		super(x, y, 32, 32, santeMax, nbPiecesDOr, vitesse,
 		        Creature.TYPE_TERRIENNE, IMAGE, "Grande Flame");
