+++ trunk/src/models/creatures/Nuage.java
@@ -29,7 +29,7 @@ public class Nuage extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public Nuage(int santeMax, int nbPiecesDOr, double vitesse)
+	public Nuage(long santeMax, int nbPiecesDOr, double vitesse)
 	{
 		this(0, 0, santeMax, nbPiecesDOr,vitesse);
 	}
@@ -43,7 +43,7 @@ public class Nuage extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public Nuage(int x, int y, int santeMax, int nbPiecesDOr, double vitesse)
+	public Nuage(int x, int y, long santeMax, int nbPiecesDOr, double vitesse)
 	{
 		super(x, y, 14, 14, santeMax, nbPiecesDOr, vitesse, 
 		      Creature.TYPE_AERIENNE, IMAGE, "Nuage");
